drm/panel/samsung-sofef00: Add panel for OnePlus 6/T devices
The OnePlus 6/T devices use different panels however they are
functionally identical with much of the commands being shared. The
panels don't appear to be used by any other devices so combine them as
one driver that is specific to the devices.

The panels are: samsung,sofef00
and             samsung,s6e3fc2x01

Signed-off-by: Caleb Connolly <caleb@connolly.tech>