drm/panel/samsung-sofef00: fix s6e3fc2x01 support
The s6e3fc2x01 doesn't seem to work properly with this driver, if it
ever goes into reset it's not possible to initialise it again.

This patch is a very hacky solution to just not reset the panel, testing
shows no change in behaviour between the panels with this patch applied,
both are able to turn off/on properly and function as expected.

This doesn't appear to be an init sequence problem, I've tested a lot of
different approaches but I can't seem to get the panel to turn on if it
gets reset.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>