rpmsg: core: Only unregister devices that belong to &rpmsg_bus
Typically the RPMSG drivers register all sub-devices under one
parent device (e.g. smd-edge for qcom_smd). And in general there
should be only RPMSG devices directly below such a parent device.

However, for the WWAN subsystem we need to create a "wwan" subdevice
below a shared parent device. For example, consider the following
setup on some Qualcomm SoCs. There are two RPMSG channels provided
by a remoteproc:

  - DATA5_CNTL (QMI)
  - DATA4 (AT)

These two channels will probe as two separate RPMSG sub-devices,
and will then separately register a WWAN port. To make them show
up as a single WWAN device with muiltiple ports, the WWAN device
must be created on a common parent device, in this case the
"smd-edge" where all rpmsg subdevices are created.

This works fine in general, and the WWAN subsystem will take care
of unregistering that device when all WWAN ports are unregistered.
But at the moment the RPMSG drivers forcibly unregister *all* devices
when the driver is removed, no matter if it belongs to the rpmsg bus
or not. This means that the WWAN devices are unregistered twice,
which causes strange warnings/crashes and prevents the kernel from
shutting down.

  sysfs group 'power' not found for kobject 'wwan0p1AT'
  WARNING: CPU: 3 PID: 1556 at fs/sysfs/group.c:279 sysfs_remove_group+0x98/0xac
  sysfs group 'power' not found for kobject 'wwan0p2QMI'
  WARNING: CPU: 3 PID: 1556 at fs/sysfs/group.c:279 sysfs_remove_group+0x98/0xac
  Unable to handle kernel NULL pointer dereference at virtual address 0000000000000070
  Internal error: Oops: 96000004 [#1] PREEMPT SMP
  Call trace:
   kernfs_find_ns+0x28/0x130
   ...
   device_unregister+0x24/0x7c
   wwan_remove_port+0xa8/0xf0
   rpmsg_wwan_ctrl_remove+0x1c/0x30
   rpmsg_dev_remove+0x48/0x90
   ...
   qcom_smd_unregister_edge+0x44/0x84

This can be avoided easily by checking that the devices we try to
unregister actually belong to the &rpmsg_bus.

Also, checking this in rpmsg_device_match() will prevent accessing
invalid memory there which usually causes the wrong device to be
returned.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>